+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [0.9.13] - 2021-12-2
 ### Added
 ### Changed
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CASMNET-1054 - make sure dhcp-helper does not patch in an IP when there is already an IP outside of loading BSS data
 - CASMNET-1055 - make sure xname is used when trying to repair SMD data after node move
 ### Security
+
 
 ## [0.9.12] - 2021-11-30
 ### Added